Add head-only formula for neatvi
neatvi is a small ex/vi-like editor that supports UTF8 as well as bidirectional
text.

By default, `make` creates an executable called `vi`, but since this is still
a work in progress rather than a full replacement for `vi` or `vim`, I've
written the formula to install it as `neatvi`. That way users can play with it
without a larger commitment.

Closes #143.

Signed-off-by: Dominyk Tiller <dominyktiller@gmail.com>